+++ selftest-core/src/test/java/net/oneandone/httpselftest/test/api/TestConfigsTest.java
@@ -11,6 +11,7 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 
+import org.junit.Before;
 import org.junit.Test;
 
 import net.oneandone.httpselftest.test.api.TestConfigs;
@@ -18,6 +19,15 @@ import net.oneandone.httpselftest.test.api.TestConfigs.Values;
 
 @SuppressWarnings("unused")
 public class TestConfigsTest {
+
+    private static final String NOT_NULL = "must not be null";
+
+    private TestConfigs p1Config;
+
+    @Before
+    public void before() {
+        p1Config = new TestConfigs("p1");
+    }
 
     @Test
     public void new_testconfigs() {
@@ -27,13 +37,13 @@ public class TestConfigsTest {
     }
 
     @Test
-    public void new_testconfigs_null_array() {
-        assertThatThrownBy(() -> new TestConfigs((String[]) null)).hasMessageContaining("must not be null");
+    public void new_testconfigs_nullArray() {
+        assertThatThrownBy(() -> new TestConfigs((String[]) null)).hasMessageContaining(NOT_NULL);
     }
 
     @Test
-    public void new_testconfigs_null_null() {
-        assertThatThrownBy(() -> new TestConfigs(null, null)).hasMessageContaining("must not be null");
+    public void new_testconfigs_nullnull() {
+        assertThatThrownBy(() -> new TestConfigs(null, null)).hasMessageContaining(NOT_NULL);
     }
 
     @Test
@@ -42,28 +52,72 @@ public class TestConfigsTest {
     }
 
     @Test
+    public void new_testconfigs_emptyString() {
+        assertThatThrownBy(() -> new TestConfigs("")).hasMessageContaining("must not be empty");
+    }
+
+    @Test
+    public void fixed_nullArray() throws Exception {
+        assertThatThrownBy(() -> p1Config.fixed((String[]) null)).hasMessageContaining(NOT_NULL);
+    }
+
+    @Test
+    public void fixed_nullnull() throws Exception {
+        assertThatThrownBy(() -> p1Config.fixed(null, null)).hasMessageContaining(NOT_NULL);
+    }
+
+    @Test
+    public void fixed_empty() throws Exception {
+        assertThatThrownBy(() -> p1Config.fixed()).hasMessageContaining("must not be empty");
+    }
+
+    @Test
+    public void fixed_emptyString() throws Exception {
+        assertThatThrownBy(() -> p1Config.fixed("")).hasMessageContaining("must not be empty");
+    }
+
+    @Test
+    public void fixed_duplicate() throws Exception {
+        assertThatThrownBy(() -> p1Config.fixed("p1", "p1")).hasMessageContaining("must be unique");
+    }
+
+    @Test
+    public void fixed_unknownParameter() throws Exception {
+        assertThatThrownBy(() -> p1Config.fixed("p2")).hasMessageContaining("parameter unknown");
+    }
+
+    @Test
+    public void fixed_twice() throws Exception {
+        p1Config.fixed("p1");
+        assertThatThrownBy(() -> p1Config.fixed("p1")).hasMessageContaining("already set");
+    }
+
+    @Test
+    public void fixed_afterPut() throws Exception {
+        p1Config.put("c1", "v1");
+        assertThatThrownBy(() -> p1Config.fixed("p1")).hasMessageContaining("before adding values");
+    }
+
+    @Test
     public void put_nullId() {
-        TestConfigs c = new TestConfigs("a");
-        assertThatThrownBy(() -> c.put(null, "va")).hasMessageContaining("must not be null");
+        assertThatThrownBy(() -> p1Config.put(null, "v1")).hasMessageContaining(NOT_NULL);
     }
 
     @Test
     public void put_nullArray() {
-        TestConfigs c = new TestConfigs("p1");
-        assertThatThrownBy(() -> c.put("c1", (String[]) null)).hasMessageContaining("must not be null");
+        assertThatThrownBy(() -> p1Config.put("c1", (String[]) null)).hasMessageContaining(NOT_NULL);
     }
 
     @Test
     public void put_nullValue() {
         TestConfigs c = new TestConfigs("p1", "p2");
-        assertThatThrownBy(() -> c.put("c1", "v1", null)).hasMessageContaining("must not be null").hasMessageContaining("'p2'");
+        assertThatThrownBy(() -> c.put("c1", "v1", null)).hasMessageContaining(NOT_NULL).hasMessageContaining("'p2'");
     }
 
     @Test
     public void put_duplicateId() {
-        TestConfigs c = new TestConfigs("a");
-        c.put("a", "va");
-        assertThatThrownBy(() -> c.put("a", "va")).hasMessageContaining("already exists");
+        p1Config.put("p1", "v1");
+        assertThatThrownBy(() -> p1Config.put("p1", "v1")).hasMessageContaining("already exists");
     }
 
     @Test
@@ -79,70 +133,130 @@ public class TestConfigsTest {
     }
 
     @Test
-    public void testconfigs_isEmpty() {
-        TestConfigs c = new TestConfigs("a");
-        assertThat(c.isEmpty()).as("should be empty").isTrue();
-        c.put("a", "va");
-        assertThat(c.isEmpty()).as("should not be empty").isFalse();
+    public void isEmpty() {
+        assertThat(p1Config.isEmpty()).as("should be empty").isTrue();
+        p1Config.put("p1", "v1");
+        assertThat(p1Config.isEmpty()).as("should not be empty").isFalse();
     }
 
     @Test
-    public void testconfigs_getNull() {
-        TestConfigs c = new TestConfigs("a");
-        assertThatThrownBy(() -> c.getValues(null)).hasMessageContaining("must not be null");
+    public void getValues_null() {
+        assertThatThrownBy(() -> p1Config.getValues(null)).hasMessageContaining(NOT_NULL);
     }
 
     @Test
-    public void testconfigs_get_exists() {
-        TestConfigs c = new TestConfigs("p1");
-        c.put("c1", "v1");
-        c.getValues("c1");
+    public void getValues_exists() {
+        p1Config.put("c1", "v1");
+        p1Config.getValues("c1");
     }
 
     @Test
-    public void testconfigs_get_notexists() {
-        TestConfigs c = new TestConfigs("p1");
-        c.put("c1", "v1");
-
-        assertThatThrownBy(() -> c.getValues("c2")).hasMessageContaining("not found");
+    public void getValues_notExists() {
+        p1Config.put("c1", "v1");
+        assertThatThrownBy(() -> p1Config.getValues("c2")).hasMessageContaining("not found");
     }
 
     @Test
     public void getIds() {
-        TestConfigs c = new TestConfigs("p1");
-        c.put("c1", "v1");
-        c.put("c2", "v2");
-        c.put("c3", "v3");
-        assertThat(c.getIds()).isEqualTo((new HashSet<String>(asList("c1", "c2", "c3"))));
+        p1Config.put("c1", "v1");
+        p1Config.put("c2", "v2");
+        p1Config.put("c3", "v3");
+        assertThat(p1Config.getIds()).isEqualTo((new HashSet<String>(asList("c1", "c2", "c3"))));
     }
 
     @Test
-    public void getConfigId() throws Exception {
-        TestConfigs c = new TestConfigs("p1");
-        c.put("c1", "v1");
-        c.put("c2", "v2");
-        assertThat(c.getValues("c1").activeConfigId()).containsSame("c1");
-        assertThat(c.getValues("c2").activeConfigId()).containsSame("c2");
+    public void activeConfigId() throws Exception {
+        p1Config.put("c1", "v1");
+        p1Config.put("c2", "v2");
+        assertThat(p1Config.getValues("c1").activeConfigId()).containsSame("c1");
+        assertThat(p1Config.getValues("c2").activeConfigId()).containsSame("c2");
+    }
+
+    @Test
+    public void create_nullMap() throws Exception {
+        assertThatThrownBy(() -> p1Config.create(null)).hasMessageContaining(NOT_NULL);
+    }
+
+    @Test
+    public void create_null_value() {
+        Map<String, String> params = new HashMap<>();
+        params.put("p1", null);
+        assertThatThrownBy(() -> p1Config.create(params)).hasMessageContaining(NOT_NULL).hasMessageContaining("'p1'");
+    }
+
+    @Test
+    public void create_unknownParameter_isIgnored() {
+        Map<String, String> params = new HashMap<>();
+        params.put("p2", "v2");
+        p1Config.create(params);
+    }
+
+    @Test
+    public void create_missingParameter_usesDefaultValue() {
+        TestConfigs c = new TestConfigs("p1", "p2");
+        Map<String, String> params = new HashMap<>();
+        params.put("p2", "v2");
+        assertThat(c.create(params).get("p1")).isEqualTo("");
+    }
+
+    @Test
+    public void create_fixed() throws Exception {
+        String FIXED = "f2";
+        TestConfigs c = new TestConfigs("p1", FIXED);
+        c.fixed(FIXED);
+        c.put("c1", "v1", "fixed_value");
+
+        HashMap<String, String> params = new HashMap<>();
+        params.put("p1", "v1");
+        params.put(FIXED, "v2");
+
+        Values v;
+        v = c.create(params);
+        assertThat(v.get("p1")).isEqualTo("v1");
+        assertThat(v.get(FIXED)).isEqualTo("");
+
+        v = c.create("c1", params);
+        assertThat(v.get("p1")).isEqualTo("v1");
+        assertThat(v.get(FIXED)).isEqualTo("fixed_value");
+        assertThat(v.getFixedParameterNames()).containsExactly(FIXED);
+    }
+
+    @Test
+    public void create_ok() {
+        TestConfigs c = new TestConfigs("p1", "p2");
+        Map<String, String> params = new HashMap<>();
+        params.put("p1", "v1");
+        params.put("p2", "v2");
+        c.create(params);
     }
 
     @Test
     public void createWithId_nullId() {
-        TestConfigs c = new TestConfigs("p1");
-        assertThatThrownBy(() -> c.create(null, emptyMap())).hasMessageContaining("must not be null");
+        assertThatThrownBy(() -> p1Config.create(null, emptyMap())).hasMessageContaining(NOT_NULL);
+    }
+
+    @Test
+    public void createWithId_nullParams() {
+        p1Config.put("c1", "v1");
+        assertThatThrownBy(() -> p1Config.create("c1", null)).hasMessageContaining(NOT_NULL);
     }
 
     @Test
     public void createWithId_unknownId() throws Exception {
-        TestConfigs c = new TestConfigs("p1");
-        c.put("c1", "v1");
-        assertThatThrownBy(() -> c.create("c2", emptyMap())).hasMessageContaining("Unknown config id");
+        p1Config.put("c1", "v1");
+        assertThatThrownBy(() -> p1Config.create("c2", emptyMap())).hasMessageContaining("Unknown config id");
     }
 
     @Test
-    public void createWithId() throws Exception {
-        TestConfigs c = new TestConfigs("p1");
-        c.put("c1", "v1");
-        c.create("c1", emptyMap());
+    public void createWithId_ok() throws Exception {
+        p1Config.put("c1", "v1");
+        p1Config.create("c1", emptyMap());
+    }
+
+    @Test
+    public void createEmpty() {
+        Values v = p1Config.createEmpty();
+        assertThat(v.get("p1")).isEqualTo("");
     }
 
     @Test
@@ -166,65 +280,17 @@ public class TestConfigsTest {
 
     @Test
     public void values_get_unknownParameter() {
-        TestConfigs c = new TestConfigs("p1");
-        c.put("c1", "v1");
-        Values values = c.getValues("c1");
+        p1Config.put("c1", "v1");
+        Values values = p1Config.getValues("c1");
         assertThatThrownBy(() -> values.get("p2")).hasMessageContaining("Unknown parameter");
 
     }
 
     @Test
     public void values_get_null() {
-        TestConfigs c = new TestConfigs("p1");
-        c.put("c1", "v1");
-        Values values = c.getValues("c1");
+        p1Config.put("c1", "v1");
+        Values values = p1Config.getValues("c1");
         assertThatThrownBy(() -> values.get(null)).hasMessageContaining("Unknown parameter name");
-    }
-
-    @Test
-    public void createEmpty() {
-        TestConfigs c = new TestConfigs("p1");
-        Values v = c.createEmpty();
-        assertThat(v.get("p1")).isEqualTo("");
-    }
-
-    @Test
-    public void create_ok() {
-        TestConfigs c = new TestConfigs("p1", "p2");
-        Map<String, String> params = new HashMap<>();
-        params.put("p1", "v1");
-        params.put("p2", "v2");
-        c.create(params);
-    }
-
-    @Test
-    public void create_nullMap() throws Exception {
-        TestConfigs c = new TestConfigs("p1");
-        assertThatThrownBy(() -> c.create(null)).hasMessageContaining("must not be null");
-    }
-
-    @Test
-    public void create_null_value() {
-        TestConfigs c = new TestConfigs("p1");
-        Map<String, String> params = new HashMap<>();
-        params.put("p1", null);
-        assertThatThrownBy(() -> c.create(params)).hasMessageContaining("must not be null").hasMessageContaining("'p1'");
-    }
-
-    @Test
-    public void create_unknownParameter_isIgnored() {
-        TestConfigs c = new TestConfigs("p1");
-        Map<String, String> params = new HashMap<>();
-        params.put("p2", "v2");
-        c.create(params);
-    }
-
-    @Test
-    public void create_missingParameter_usesDefaultValue() {
-        TestConfigs c = new TestConfigs("p1", "p2");
-        Map<String, String> params = new HashMap<>();
-        params.put("p2", "v2");
-        assertThat(c.create(params).get("p1")).isEqualTo("");
     }
 
 }