+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 <!-- Categories: Added, Changed, Deprecated, Removed, Fixed, Security -->
 
 ## [Unreleased]
+- added support for unmodifiable test-provided parameters via `TestConfigs.fixed(..)`
 
 ## [0.2.2] - 2019-03-28
 - fix artifact deployment to sonatype